+++ src/main/java/com/circle/circle_backend/circle/controller/CircleReadController.java
@@ -1,6 +1,6 @@
 package com.circle.circle_backend.circle.controller;
 
-import com.circle.circle_backend.circle.controller.port.CircleResponse;
+import com.circle.circle_backend.circle.dto.response.CircleResponse;
 import com.circle.circle_backend.circle.controller.port.CircleService;
 import com.circle.circle_backend.circle.domain.Circle;
 import com.circle.circle_backend.circle.domain.enums.Category;
