+++ src/main/java/com/circle/circle_backend/circle/controller/port/CircleService.java
@@ -2,8 +2,8 @@ package com.circle.circle_backend.circle.controller.port;
 
 import com.circle.circle_backend.circle.domain.Circle;
 import com.circle.circle_backend.circle.domain.enums.Category;
-import com.circle.circle_backend.circle.domain.port.CircleCreateRequest;
-import com.circle.circle_backend.circle.domain.port.CircleUpdateRequest;
+import com.circle.circle_backend.circle.dto.request.CircleCreateRequest;
+import com.circle.circle_backend.circle.dto.request.CircleUpdateRequest;
 import com.circle.circle_backend.user.domain.User;
 
 import java.util.List;
