+++ src/main/java/com/circle/circle_backend/circle/dto/request/CircleUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.circle.circle_backend.circle.domain.port;
+package com.circle.circle_backend.circle.dto.request;
 
 import com.circle.circle_backend.circle.domain.enums.Category;
 import com.circle.circle_backend.circle.domain.enums.Characteristic;
