+++ src/main/java/com/circle/circle_backend/circle/service/CircleServiceImpl.java
@@ -3,10 +3,10 @@ package com.circle.circle_backend.circle.service;
 import com.circle.circle_backend.circleMember.domain.CircleMember;
 import com.circle.circle_backend.circle.domain.enums.Category;
 import com.circle.circle_backend.circleMember.domain.enums.UserRole;
-import com.circle.circle_backend.circle.domain.port.CircleCreateRequest;
+import com.circle.circle_backend.circle.dto.request.CircleCreateRequest;
 import com.circle.circle_backend.circle.controller.port.CircleService;
 import com.circle.circle_backend.circle.domain.Circle;
-import com.circle.circle_backend.circle.domain.port.CircleUpdateRequest;
+import com.circle.circle_backend.circle.dto.request.CircleUpdateRequest;
 import com.circle.circle_backend.circle.infrastructure.entity.CircleEntity;
 import com.circle.circle_backend.circleMember.service.CircleMemberRepository;
 import com.circle.circle_backend.common.domain.ResourceNotFoundException;
