+++ src/main/java/com/circle/circle_backend/enrollment/controller/EnrollmentCreateController.java
@@ -1,6 +1,7 @@
 package com.circle.circle_backend.enrollment.controller;
 
-import com.circle.circle_backend.enrollment.controller.port.EnrollmentResponse;
+import com.circle.circle_backend.enrollment.controller.dto.response.EnrollmentResponse;
+import com.circle.circle_backend.enrollment.controller.port.EnrollmentService;
 import com.circle.circle_backend.enrollment.domain.Enrollment;
 import com.circle.circle_backend.enrollment.domain.enums.EnrollmentState;
 import com.circle.circle_backend.security.service.UserDetailsImpl;
