+++ src/main/java/com/circle/circle_backend/user/controller/UserCreateController.java
@@ -1,7 +1,9 @@
 package com.circle.circle_backend.user.controller;
 
+import com.circle.circle_backend.user.controller.port.UserService;
 import com.circle.circle_backend.user.domain.User;
-import com.circle.circle_backend.user.domain.UserCreateRequest;
+import com.circle.circle_backend.user.dto.request.UserCreateRequest;
+import com.circle.circle_backend.user.dto.response.UserResponse;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
