+++ src/main/java/com/circle/circle_backend/user/controller/UserPatchController.java
@@ -1,8 +1,10 @@
 package com.circle.circle_backend.user.controller;
 
 import com.circle.circle_backend.security.service.UserDetailsImpl;
+import com.circle.circle_backend.user.controller.port.UserService;
 import com.circle.circle_backend.user.domain.User;
-import com.circle.circle_backend.user.domain.UserPatchRequest;
+import com.circle.circle_backend.user.dto.request.UserPatchRequest;
+import com.circle.circle_backend.user.dto.response.MyInfoResponse;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +21,7 @@ public class UserPatchController {
 
     @PatchMapping
     public ResponseEntity<MyInfoResponse> patch(@AuthenticationPrincipal UserDetailsImpl userDetails,
-                                              @RequestBody UserPatchRequest userPatchRequest) {
+                                                @RequestBody UserPatchRequest userPatchRequest) {
         User user = userService.patch(userDetails.getUser(), userPatchRequest);
         return ResponseEntity.ok(MyInfoResponse.from(user));
     }
