+++ src/main/java/com/circle/circle_backend/user/dto/response/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.circle.circle_backend.user.controller;
+package com.circle.circle_backend.user.dto.response;
 
 import com.circle.circle_backend.user.domain.User;
 import com.circle.circle_backend.user.domain.enums.Language;
