+++ src/main/java/com/circle/circle_backend/user/service/UserServiceImpl.java
@@ -2,10 +2,10 @@ package com.circle.circle_backend.user.service;
 
 import com.circle.circle_backend.common.domain.ResourceNotFoundException;
 import com.circle.circle_backend.security.utils.PasswordUtils;
-import com.circle.circle_backend.user.controller.UserService;
+import com.circle.circle_backend.user.controller.port.UserService;
 import com.circle.circle_backend.user.domain.User;
-import com.circle.circle_backend.user.domain.UserCreateRequest;
-import com.circle.circle_backend.user.domain.UserPatchRequest;
+import com.circle.circle_backend.user.dto.request.UserCreateRequest;
+import com.circle.circle_backend.user.dto.request.UserPatchRequest;
 import com.circle.circle_backend.user.infrastructure.entity.UserEntity;
 import jakarta.transaction.Transactional;
 import lombok.Builder;
